charon-fsm: Add multiple tunnel support
Multiple IKE SAs will cause distinct tunnel interfaces
to be created, so a custom updown plugin is needed. This
handles sending the correct interface name for the given
IKE SA to the updown script. It is based on the default updown
plugin.

The fsm_updown script is also updated to use the passed in
PLUTO_INTERFACE environment variable instead of hardcoding
the tunnel interface name. Additionally, it now sends
the PLUTO_PEER and PLUTO_INTERFACE to runBaseConfig and
runBaseConfigforV6 as parameters.

CRs-Fixed: 1029270
Change-Id: Id480cfd9f9ef8d5bcfbc9b2fda987e06b4188843
Signed-off-by: Sivanesan Rajapupathi <srajap@codeaurora.org>