kernel-pfroute: Maintain virtual flag when repopulating interface addrs
When adding a virtual IP on a TUN interface, the interface might get
activated after we already set the virtual flag for the added address.
As this repopulates the addresses on the interface, this cleared the
flag and the address would no longer be treated as virtual IP when
installing routes for CHILD_SAs that reference it in their local traffic
selectors.