child-sa: Handle refcount overflow for unique mark allocation gracefully
The refcount_t for allocating unique marks may overflow or hit the special
value for unique marks, in the worst case not setting a mark on CHILD_SAs
that should have one.

As (potentially two) marks are allocated only for newly created CHILD_SAs,
but not for rekeying, this not very likely. Still, if a setup uses
aggressive re-authentication and or re-creates CHILD_SAs every minute,
a gateway with 100'000 tunnels may hit the overflow within a month uptime.
So prevent allocation of invalid marks by using a new helper function.